server/requirements.txt
asgiref==3.2.10
biopython==1.77
boto3==1.14.18
botocore==1.17.18
Django==3.0.8
django-cors-headers==3.4.0
docutils==0.15.2
jmespath==0.10.0
numpy==1.19.0
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
